Changed folder structure - FrontEnd
