+++ src/webapp/myreact/src/shared/Navbar.js
@@ -5,19 +5,13 @@ import Drawer from '@mui/material/Drawer';
 import ListItem from '@mui/material/ListItem';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import ListItemText from '@mui/material/ListItemText';
-import Time from './Time'
-import DropDown from './DropDown'
 import AccountCircleRoundedIcon from '@mui/icons-material/AccountCircleRounded';
 import SettingsBackupRestoreRoundedIcon from '@mui/icons-material/SettingsBackupRestoreRounded';
 import AccountBalanceWalletIcon from '@mui/icons-material/AccountBalanceWallet';
 import LiveHelpIcon from '@mui/icons-material/LiveHelp';
 import EmailIcon from '@mui/icons-material/Email';
 import LogoutIcon from '@mui/icons-material/Logout';
-import ArrowForwardIcon from '@mui/icons-material/ArrowForward';
-import Button from '@mui/material/Button';
-import TextField from '@material-ui/core/TextField';
 import { makeStyles } from '@material-ui/core/styles';
-import Grid from '@mui/material/Grid';
 
 const useStyles = makeStyles((theme) => ({
   root: {
