+++ src/webapp/myreact/src/shared/Time.js
@@ -1,0 +1,27 @@
+import * as React from 'react';
+import TextField from '@mui/material/TextField';
+import AdapterDateFns from '@mui/lab/AdapterDateFns';
+import LocalizationProvider from '@mui/lab/LocalizationProvider';
+import DateTimePicker from '@mui/lab/DateTimePicker';
+import { Grid } from '@mui/material';
+
+export default function MaterialUIPickers() {
+  const [value, setValue] = React.useState(new Date());
+
+  const handleChange = (newValue) => {
+    setValue(newValue);
+  };
+
+  return (
+    <LocalizationProvider dateAdapter={AdapterDateFns}>
+      <Grid>
+      <DateTimePicker
+          label="Schedule Ride"
+          value={value}
+          onChange={handleChange}
+          renderInput={(params) => <TextField {...params} />}
+        />
+      </Grid>
+    </LocalizationProvider>
+  );
+}
